data_analysis_tool/requirements_data_tool.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
lxml>=5.0.0
httpx>=0.26.0
sse-starlette>=1.8.0
python-pptx>=0.6.23
python-docx>=1.1.0
reportlab>=4.0.0
matplotlib>=3.8.0
